@@ -1,10 +1,12 @@
-﻿using Microsoft.AspNetCore.Identity;
+﻿// Copyright (c) Duende Software. All rights reserved.
+// See LICENSE in the project root for license information.
 
-namespace FreeCourse.IdentityServer.Models
+
+using Microsoft.AspNetCore.Identity;
+
+namespace FreeCourse.IdentityServer.Models;
+
+// Add profile data for application users by adding properties to the ApplicationUser class
+public class ApplicationUser : IdentityUser
 {
-    // Add profile data for application users by adding properties to the ApplicationUser class
-    public class ApplicationUser : IdentityUser
-    {
-        public string City { get; set; }
-    }
 }
